C adding colors into chessboard
