+++ src/Bishop.js
@@ -13,32 +13,32 @@ var __extends = (this && this.__extends) || (function () {
     };
 })();
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.Rook = void 0;
-var PieceClass_1 = require("./PieceClass");
+exports.Bishop = void 0;
+var Piece_1 = require("./Piece");
 var logic_1 = require("./logic");
 var logic_2 = require("./logic");
-var Rook = /** @class */ (function (_super) {
-    __extends(Rook, _super);
-    function Rook(color, location) {
-        return _super.call(this, 'rock', color, location) || this;
+var Bishop = /** @class */ (function (_super) {
+    __extends(Bishop, _super);
+    function Bishop(color, location) {
+        return _super.call(this, 'bishop', color, location) || this;
     }
-    Rook.prototype.checkPossibleMoves = function () {
+    Bishop.prototype.checkPossibleMoves = function () {
         this.check = false;
         this.possibleLocations = [];
         for (var i = 0; i < 4; i++) {
             var dir = [0, 0];
             switch (i) {
                 case 0:
-                    dir = [1, 0];
+                    dir = [1, 1];
                     break;
                 case 1:
-                    dir = [-1, 0];
+                    dir = [-1, 1];
                     break;
                 case 2:
-                    dir = [0, 1];
+                    dir = [1, -1];
                     break;
                 case 3:
-                    dir = [0, -1];
+                    dir = [-1, -1];
                     break;
             }
             for (var j = 1; j < 8; j++) {
@@ -49,9 +49,9 @@ var Rook = /** @class */ (function (_super) {
                     checkingPosition[1] > 8)
                     break;
                 var pieceOnWay = logic_2.AREASARRAY[logic_1.getAreaArrayIndex(checkingPosition)].piece;
-                if (pieceOnWay instanceof PieceClass_1.Piece && pieceOnWay.color == this.color)
+                if (pieceOnWay instanceof Piece_1.Piece && pieceOnWay.color == this.color)
                     break;
-                else if (pieceOnWay instanceof PieceClass_1.Piece && pieceOnWay.color !== this.color) {
+                else if (pieceOnWay instanceof Piece_1.Piece && pieceOnWay.color !== this.color) {
                     if (pieceOnWay.type == 'king')
                         this.check = true;
                     this.possibleLocations.push(checkingPosition);
@@ -61,6 +61,6 @@ var Rook = /** @class */ (function (_super) {
             }
         }
     };
-    return Rook;
-}(PieceClass_1.Piece));
-exports.Rook = Rook;
+    return Bishop;
+}(Piece_1.Piece));
+exports.Bishop = Bishop;
