+++ src/chessboard.ts
@@ -23,7 +23,6 @@ outerBoard.appendChild(letters);
 outerBoard.appendChild(board);
 
 export function chessboard() {
-
     function getBoard() {
         for (let i = 8; i >= 1; i--) {
             let row = document.createElement('div');
@@ -42,9 +41,9 @@ export function chessboard() {
                 }
                 square.classList.add('square');
                 if (j % 2 === 0) {
-                    square.style.backgroundColor = '#6C1A31';
+                    square.classList.add('firstColor');
                 } else {
-                    square.style.backgroundColor = '#D67E03';
+                    square.classList.add('secondColor');
                 }
                 row.appendChild(square);
             }
